requirements.txt
django
django-rest-framework
django-cors-headers
drf-yasg
nltk
numpy
gensim
tensorflow